created the supercharger.optimize package.
